Модульные тесты [#505277]

+++ src/main/java/ru/job4j/Converter.java
@@ -2,12 +2,12 @@ package ru.job4j;
 
 public class Converter {
 
-    public static float rubleToEuro(float value) {
+    public static double rubleToEuro(double value) {
         // Assuming the conversion rate is 1 Euro = 90 Rubles
         return value / 90;
     }
 
-    public static float rubleToDollar(float value) {
+    public static double rubleToDollar(double value) {
         // Assuming the conversion rate is 1 Dollar = 75 Rubles
         return value / 75;
     }
